+++ R/utils.R
@@ -5,4 +5,9 @@
 NULL
 
 # To avoid R CMD check notes about undefined global variables
-utils::globalVariables(c("x_names", "rel_imp", "object", "new_data", "nn_module", "fit_class", "epoch", "loss", "set"))
+utils::globalVariables(c(
+    "x_names", "rel_imp", "object", "new_data", "nn_module", "fit_class",
+    "epoch", "loss", "set",
+    "fitted_val", "residuals", "actual_val",
+    "actual", "predicted", "n", "prop"
+))